Rename file itself of ChatTitlesListCache
